Move graph class in separated files
